:tada: feat: able to set departments for 10 codes
